backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
celery==5.3.6
redis==5.0.4
motor==3.3.2
pymongo==4.6.1
python-dotenv==1.0.1
python-multipart==0.0.9